+++ docs/ProcessManager.Orchestrations.Abstractions/ReleaseNotes/ReleaseNotes.md
@@ -1,8 +1,12 @@
 # ProcessManager.Orchestrations.Abstractions Release Notes
 
-## Version 0.13.0
+## Version 0.14.0
 
 - Add `DataHubTypes` derived types to be shared with consumer.
+
+## Version 0.13.0
+
+- Dependent NuGet packages updated
 
 ## Version 0.12.2